experiments/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy<2
sentence-transformers
torch
httpx
qdrant-client